forum

+++ app/dashboard/forum/forum.tsx
@@ -9,6 +9,7 @@ import "react-markdown-editor-lite/lib/index.css";
 import { Tab, Tabs, TabList, TabPanel } from "react-tabs";
 import "react-tabs/style/react-tabs.css";
 import "./markdown-styles.css";
+import ConfirmDeletionSectionForm from "@/components/confirmDeletionModalSection";
 import { CirclePlus } from "lucide-react";
 import { toast } from "react-toastify";
 import {
@@ -21,6 +22,7 @@ import {
     deleteForumPosts,
     getForumPosts,
 } from "./actions";
+import RenderPosts from "@/components/forum/renderPosts";
 
 interface ForumProps {
     readonly userEmail: string;
@@ -62,6 +64,9 @@ export default function Forum({ userEmail, plan, access_level }: ForumProps) {
     const [postTitle, setPostTitle] = useState("");
     const [editingTable, setEditingTable] = useState<ForumTable | null>(null);
 
+    const [deleteModalVisible, setDeleteModalVisible] = useState(false);
+    const [tableToDelete, setTableToDelete] = useState<string | null>(null);
+
 
     useEffect(() => {
         fetchTables();
@@ -86,14 +91,22 @@ export default function Forum({ userEmail, plan, access_level }: ForumProps) {
         }
     };
 
-    const handleDeleteTable = async (tableId: string) => {
+    const openDeleteModal = (tableId: string) => {
+        setTableToDelete(tableId);
+        setDeleteModalVisible(true);
+    };
+
+    const handleDeleteTable = async () => {
         if (confirm("Are you sure you want to delete this table? This action cannot be undone.")) {
             try {
+                if (!tableToDelete) {
+                    toast.error("Table ID is missing.");
+                    return;
+                }
                 const formData = new FormData();
-                formData.append("id", tableId);
+                formData.append("id", tableToDelete);
                 await deleteForumTable({ message: "" }, formData);
                 fetchTables(); // Refresh the table list
-                toast.success("Table deleted successfully.");
             } catch (error) {
                 toast.error("Failed to delete table.");
             }
@@ -113,7 +126,6 @@ export default function Forum({ userEmail, plan, access_level }: ForumProps) {
             fetchPosts();
             setMarkdown("");
             setActiveEditor(null);
-            toast.success("Post submitted successfully.");
         } catch (error) {
             toast.error("Failed to submit post.");
         }
@@ -152,7 +164,6 @@ export default function Forum({ userEmail, plan, access_level }: ForumProps) {
             setMarkdown("");
             setPostTitle("");
             setActiveEditor(null);
-            toast.success("Post updated successfully.");
         } catch (error) {
             toast.error("Failed to update post.");
         }
@@ -164,7 +175,6 @@ export default function Forum({ userEmail, plan, access_level }: ForumProps) {
         try {
             await deleteForumPosts({ message: "" }, formData);
             fetchPosts();
-            toast.success("Post deleted successfully.");
         } catch (error) {
             toast.error("Failed to delete post.");
         }
@@ -186,11 +196,9 @@ export default function Forum({ userEmail, plan, access_level }: ForumProps) {
                 // Update existing table
                 formData.append("id", editingTable.id);
                 await updateForumTable(null, formData);
-                toast.success("Table updated successfully.");
             } else {
                 // Add new table
                 await insertForumTable(null, formData);
-                toast.success("Table added successfully.");
             }
             fetchTables();
             resetForm();
@@ -199,18 +207,6 @@ export default function Forum({ userEmail, plan, access_level }: ForumProps) {
         }
     };
 
-
-    const deleteTable = async (tableId: string) => {
-        const formData = new FormData();
-        formData.append("id", tableId);
-        try {
-            await deleteForumTable({ message: "" }, formData);
-            fetchTables();
-            toast.success("Table deleted successfully.");
-        } catch (error) {
-            toast.error("Failed to delete table.");
-        }
-    };
 
     const likePost = async (postId: string) => {
         const post = posts.find((p) => p.id === postId);
@@ -230,7 +226,6 @@ export default function Forum({ userEmail, plan, access_level }: ForumProps) {
             try {
                 await updateForumPosts(null, formData);
                 fetchPosts();
-                toast.success("Post liked.");
             } catch (error) {
                 toast.error("Failed to like post.");
             }
@@ -245,90 +240,16 @@ export default function Forum({ userEmail, plan, access_level }: ForumProps) {
         setMarkdown(text);
     };
 
-    const renderPosts = (parentId: string | null, tableId: string) => {
-        return posts
-            .filter((post) => post.parent_id === parentId && post.table_id === tableId).sort((a, b) => a.created_at > b.created_at ? -1 : 1)
-            .map((post) => (
-                <div
-                    key={post.id}
-                    className={`mt-4 p-4 border rounded-lg ${parentId ? "ml-8" : "ml-0"} shadow-md bg-white`}
-                >
-                    <h3 className="text-lg font-semibold">{post.title}</h3>
-                    <p className="text-sm text-gray-600">
-                        <strong>By:</strong> {post.user_name} - {post.created_at.toLocaleString()}
-                    </p>
-                    <div className="prose">
-                        <ReactMarkdown remarkPlugins={[remarkGfm]} rehypePlugins={[rehypeRaw]}>
-                            {post.description}
-                        </ReactMarkdown>
-                    </div>
-                    <div className="mt-2 flex space-x-4">
-                        <button
-                            className="text-blue-500 hover:underline"
-                            onClick={() => likePost(post.id)}
-                        >
-                            Like ({post.likes})
-                        </button>
-                        <button
-                            className="text-blue-500 hover:underline"
-                            onClick={() => setActiveEditor({ type: "comment", id: post.id })}
-                        >
-                            Reply
-                        </button>
-                        {post.user_email === userEmail && (
-                            <>
-                                <button
-                                    className="text-blue-500 hover:underline"
-                                    onClick={() => editPost(post.id)}
-                                >
-                                    Edit
-                                </button>
-                                <button
-                                    className="text-red-500 hover:underline"
-                                    onClick={() => deletePost(post.id)}
-                                >
-                                    Delete
-                                </button>
-                            </>
-                        )}
-
-
-                    </div>
-
-                    {/* Render active editor below the current post */}
-                    {activeEditor?.id === post.id && (
-                        <div className="mt-4 p-4 bg-gray-50 border rounded-lg">
-                            <h4 className="text-md font-semibold mb-2">Reply or Edit Post</h4>
-                            <MdEditor
-                                value={markdown}
-                                style={{ height: "200px" }}
-                                renderHTML={(text) => (
-                                    <ReactMarkdown remarkPlugins={[remarkGfm]} rehypePlugins={[rehypeRaw]}>
-                                        {text}
-                                    </ReactMarkdown>
-                                )}
-                                onChange={handleEditorChange}
-                            />
-                            <button
-                                onClick={() =>
-                                    activeEditor.type === "edit"
-                                        ? saveEditedPost(post)
-                                        : submitPost(postTitle, tableId, post.id)
-                                }
-                                className="mt-2 px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600"
-                            >
-                                {activeEditor.type === "edit" ? "Save Changes" : "Submit Reply"}
-                            </button>
-                        </div>
-                    )}
-                    {renderPosts(post.id, tableId)}
-                </div>
-            ));
-    };
 
     return (
         <div className="container">
             <h1>Hello, {userEmail}</h1>
+
+            {access_level === "admin" && <ConfirmDeletionSectionForm title="Are you sure you whant to delete this section?" onSubmit={handleDeleteTable} isOpen={deleteModalVisible} onOpenChange={() => {
+                setDeleteModalVisible(!deleteModalVisible);
+                setTableToDelete(null);
+            }} />}
+
             {access_level === "admin" && (
                 <div className="bg-gray-100 p-6 rounded-lg shadow-md mb-4">
                     <h2 className="text-2xl font-semibold mb-4">{editingTable ? "Edit Table" : "Add a New Table"}</h2>
@@ -385,7 +306,7 @@ export default function Forum({ userEmail, plan, access_level }: ForumProps) {
 
                                         {/* Delete Button */}
                                         <button
-                                            onClick={() => handleDeleteTable(table.id)}
+                                            onClick={() => openDeleteModal(table.id)}
                                             className="px-2 py-1 text-sm bg-red-500 text-white rounded hover:bg-red-600"
                                         >
                                             Delete
@@ -436,7 +357,23 @@ export default function Forum({ userEmail, plan, access_level }: ForumProps) {
                                     </button>
                                 </div>
                             )}
-                            {renderPosts(null, table.id)}
+                            <RenderPosts
+                                parentId={null}
+                                tableId={table.id}
+                                posts={posts}
+                                userEmail={userEmail}
+                                activeEditor={activeEditor}
+                                likePost={likePost}
+                                setActiveEditor={setActiveEditor}
+                                editPost={editPost}
+                                deletePost={deletePost}
+                                markdown={markdown}
+                                handleEditorChange={handleEditorChange}
+                                saveEditedPost={saveEditedPost}
+                                submitPost={submitPost}
+                                postTitle={postTitle}
+                                access_level={access_level}
+                            />
                         </div>
                     </TabPanel>
                 ))}
